chore: refactor library parsing
